feat: move out sendingEmails methods
Note: email.ts role is to parse argument and launch theses methods
